backend/requirements.txt
asgiref==3.7.2
gunicorn==20.1.0
Django==3.2.16
django-extensions==3.2.3
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
pytz==2024.1
Pillow==9.0.0
sqlparse==0.4.4
psycopg2-binary==2.9.3
typing_extensions==4.10.0
ipython==8.18.1
webcolors==1.13
python-dotenv==1.0.1